+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # Changelog
 
-## [3.22.0] – 2025-06-23
+## [3.22.0] – 2025-06-24
 ### ✨ Added
 - **TWW S3 Catalyst Currency**
 - **Dungeon Teleports for TWW S3**
 - **Updated TOC for 11.2.0**
+- **Instant Catalyst button**
+  - Instantly catalyst an item without the waiting time
 
 ### 🐛 Fixed
 - Character frame reliably shows **Catalyst charges** again